Move BLS parsing code to the BLS test handler
